Refectoring Utils and ConstraintVariables
