+++ tools/checked-c-convert/CMakeLists.txt
@@ -11,6 +11,8 @@ add_clang_executable(checked-c-convert
   ConstraintBuilder.cpp
   PersistentSourceLoc.cpp
   Constraints.cpp
+  ConstraintVariables.cpp
+  Utils.cpp
   )
 
 target_link_libraries(checked-c-convert
